V1.0.2 Extended out core Features Fixed Login/Register Syncing Issues Fixed Preformance issues

+++ lib/Cookie.php
@@ -16,7 +16,7 @@
 
         public static function put($name, $value, $expire)
         {
-            if(setcookie($name, $value, time() + $expire, '/'))
+            if(setcookie($name, $value, (int)(time() + $expire), '/'))
                 return true;
 
             return false;
